Exception y logica de registrar turno

+++ src/main/java/com/backend/clinicaodontologica/dto/entrada/paciente/DomicilioEntradaDto.java
@@ -10,7 +10,7 @@ public class DomicilioEntradaDto {
     @NotNull(message = "El campo numero no puede ser nulo")
     @Digits(integer = 8, fraction = 0, message = "El número debe tener como máximo 8 dígitos")
     //@Pattern(regexp = "\\d{1,8}", message = "El número debe tener como máximo 8 dígitos")
-    private Integer numero;
+    private Long numero;
 
     @NotNull(message = "El campo localidad no puede ser nulo")
     @NotBlank(message = "El campo localidad no puede estar en blanco")
@@ -23,7 +23,7 @@ public class DomicilioEntradaDto {
     public DomicilioEntradaDto() {
     }
 
-    public DomicilioEntradaDto(String calle, Integer numero, String localidad, String provincia) {
+    public DomicilioEntradaDto(String calle, Long numero, String localidad, String provincia) {
         this.calle = calle;
         this.numero = numero;
         this.localidad = localidad;
@@ -38,11 +38,11 @@ public class DomicilioEntradaDto {
         this.calle = calle;
     }
 
-    public Integer getNumero() {
+    public Long getNumero() {
         return numero;
     }
 
-    public void setNumero(Integer numero) {
+    public void setNumero(Long numero) {
         this.numero = numero;
     }
